Add process of ignoring file already symbolicked

+++ setup.sh
@@ -2,7 +2,9 @@
 
 DOT_FILES=(.tmux-rename-session)
 
-for file in ${DOT_FILES[@]}
-do
-    ln -s $HOME/dotfiles/$file $HOME/$file
+for file in ${DOT_FILES[@]}; do
+  if [ -e $HOME/$file ]; then
+    continue
+  fi
+  ln -s $HOME/dotfiles/$file $HOME/$file
 done
